Rename manifest
This shouldn't be necessary, but Joomla gives an error when "Refresh Cache"
is performed in "Extension Manager: Manage" with Xmap selected.

This applies to both Joomla 2.5 and 3.x.  In 2.5, an error is shown, but
not in 3.x.  In either case, refreshing cache causes the installed version
number to be wiped out (blank).

Because the installed version number gets wiped out, a check for updates
will find an update, even though the update version is the same as the
installed verion.