Allow menu entries attached to an osmap xml view to be accessed by  adding .xml to the path with SEO
Currently when creating a joomla menu entry attached to an xml osmap view with e.g. alias sitemap, which would result in having e.g. the URL ```https://mysite.tld/sitemap``` there is no possibility to access the URL by adding a ```.xml``` suffix to the path like https://mysite.tld/sitemap.xml because the view is implicitly only available as ```.html```. Doing so leads to the error message ```view not found [name, type, prefix]: xml, xml, osmapView```.

One possibility is to add the file ```view.xml.php``` to ```components/com_osmap/views/xml``` as a 1 to 1 copy of ```view.html.php```. 
I think having ```.html``` as a suffix does not make much sense, but should be available for backward compatibility.

Allow menu entries attached to an osmap xml view to be accessed by  adding .xml to the path with SEO